rename slicable.ts types.ts
