+++ packages/atom/src/atom.ts
@@ -2,7 +2,7 @@ import * as O from "fp-ts/Option";
 import * as Op from "monocle-ts/lib/Optional"
 import { pipe } from 'fp-ts/function'
 import { Channel } from '@effection/channel';
-import { MakeSlice, AtomConfig, Slice } from './sliceable';
+import { MakeSlice, AtomConfig, Slice } from './types';
 import { Operation } from 'effection';
 import { subscribe, Subscription, SymbolSubscribable } from '@effection/subscription';
 import { unique } from './unique';
